Fixing bugs in RedBlack tree implementation

+++ src/Symbiote.Core/Hashing/IHashingProvider.cs
@@ -2,6 +2,6 @@
 {
     public interface IHashingProvider
     {
-        int Hash<T>(T value);
+        long Hash<T>(T value);
     }
 }
